# Create a BamBlockBuffer in C to eliminate writing overhead.