--- a/src/htspy/_bammodule.c
+++ b/src/htspy/_bammodule.c
@@ -365,6 +365,35 @@ PyDoc_STRVAR(BamRecord_to_bytes__doc__,
     {"to_bytes", (PyCFunction)(void(*)(void))BamRecord_to_bytes, METH_NOARGS, \
      BamRecord_to_bytes__doc__}
 
+static inline void 
+BamRecord_to_ptr(BamRecord *self, char * dest) {
+    memcpy(dest, (char *)self + BAM_PROPERTIES_STRUCT_START,
+         BAM_PROPERTIES_STRUCT_SIZE);
+    Py_ssize_t cursor = BAM_PROPERTIES_STRUCT_SIZE;
+    
+    Py_ssize_t read_name_size = PyBytes_GET_SIZE(self->read_name);
+    memcpy(dest + cursor, PyBytes_AS_STRING(self->read_name), read_name_size);
+    cursor += read_name_size;
+
+    // Terminate read_name with NULL byte
+    dest[cursor] = 0; cursor += 1;
+
+    Py_ssize_t cigar_size = PyBytes_GET_SIZE(self->cigar);
+    memcpy(dest + cursor, PyBytes_AS_STRING(self->cigar), cigar_size);
+    cursor += cigar_size;
+
+    Py_ssize_t seq_size = PyBytes_GET_SIZE(self->seq);
+    memcpy(dest + cursor, PyBytes_AS_STRING(self->seq), seq_size);
+    cursor += seq_size;
+
+    Py_ssize_t qual_size = PyBytes_GET_SIZE(self->qual);
+    memcpy(dest + cursor, PyBytes_AS_STRING(self->qual), qual_size);
+    cursor += qual_size;
+
+    Py_ssize_t tag_size = PyBytes_GET_SIZE(self->tags);
+    memcpy(dest + cursor, PyBytes_AS_STRING(self->tags), tag_size);
+}
+
 static PyObject *
 BamRecord_to_bytes(BamRecord *self, PyObject *NoArgs)
 {
@@ -374,33 +403,7 @@ BamRecord_to_bytes(BamRecord *self, PyObject *NoArgs)
         return PyErr_NoMemory();
     }
     char * bam_bytes = PyBytes_AS_STRING(ret_val);
-    
-    memcpy(bam_bytes, (char *)self + BAM_PROPERTIES_STRUCT_START,
-         BAM_PROPERTIES_STRUCT_SIZE);
-    Py_ssize_t cursor = BAM_PROPERTIES_STRUCT_SIZE;
-    
-    Py_ssize_t read_name_size = PyBytes_GET_SIZE(self->read_name);
-    memcpy(bam_bytes + cursor, PyBytes_AS_STRING(self->read_name), read_name_size);
-    cursor += read_name_size;
-
-    // Terminate read_name with NULL byte
-    bam_bytes[cursor] = 0; cursor += 1;
-
-    Py_ssize_t cigar_size = PyBytes_GET_SIZE(self->cigar);
-    memcpy(bam_bytes + cursor, PyBytes_AS_STRING(self->cigar), cigar_size);
-    cursor += cigar_size;
-
-    Py_ssize_t seq_size = PyBytes_GET_SIZE(self->seq);
-    memcpy(bam_bytes + cursor, PyBytes_AS_STRING(self->seq), seq_size);
-    cursor += seq_size;
-
-    Py_ssize_t qual_size = PyBytes_GET_SIZE(self->qual);
-    memcpy(bam_bytes + cursor, PyBytes_AS_STRING(self->qual), qual_size);
-    cursor += qual_size;
-
-    Py_ssize_t tag_size = PyBytes_GET_SIZE(self->tags);
-    memcpy(bam_bytes + cursor, PyBytes_AS_STRING(self->tags), tag_size);
-
+    BamRecord_to_ptr(self, bam_bytes);
     return ret_val;
 }
 
@@ -505,9 +508,11 @@ static PyObject * BamBlockBuffer_write(BamBlockBuffer * self, BamRecord * bam_re
     if (final_pos > self->buffersize) {
         return PyLong_FromSsize_t(0);
     }
-    
+    BamRecord_to_ptr(bam_record, self->buffer + self->pos);
+    self->pos = final_pos; 
+    return PyLong_FromSsize_t(record_size);
+}
 
-} 
 
 static PyTypeObject BamBlockBuffer_type = {
     .tp_name = "_bam.BamBlockBuffer",