--- a/src/htspy/_bammodule.c
+++ b/src/htspy/_bammodule.c
@@ -494,7 +494,7 @@ PyDoc_STRVAR(BamBlockBuffer_write_doc,
 "fit in the buffer anymore");
 
 #define BAMRECORD_TO_BYTES_METHODDEF    \
-    {"write", (PyCFunction)(void(*)(void))BamRecord_to_bytes, METH_O, \
+    {"write", (PyCFunction)(void(*)(void))BamBlockBuffer_write, METH_O, \
      BamBlockBuffer_write_doc}
 
 static PyObject * BamBlockBuffer_write(BamBlockBuffer * self, BamRecord * bam_record) {
@@ -511,6 +511,18 @@ static PyObject * BamBlockBuffer_write(BamBlockBuffer * self, BamRecord * bam_re
     BamRecord_to_ptr(bam_record, self->buffer + self->pos);
     self->pos = final_pos; 
     return PyLong_FromSsize_t(record_size);
+}
+PyDoc_STRVAR(BamBlockBuffer_reset_doc,
+"Remove all records from the buffer.");
+
+#define BAMRECORD_TO_BYTES_METHODDEF    \
+    {"reset", (PyCFunction)(void(*)(void))BamBlockBuffer_reset, METH_NOARGS, \
+     BamBlockBuffer_reset_doc}
+
+static PyObject * 
+BamBlockBuffer_reset(BamBlockBuffer * self, PyObject *Py_UNUSED(ignore)) {
+    self->pos = 0;
+    Py_RETURN_NONE;
 }