--- a/src/htspy/_bammodule.c
+++ b/src/htspy/_bammodule.c
@@ -478,14 +478,6 @@ static PyMemberDef BamBlockBuffer_members[] = {
     {NULL}
 };
 
-static Py_ssize_t 
-BamBlockBuffer__len__(BamBlockBuffer * self) {
-    return self->pos;
-}
-
-static PySequenceMethods BamBlockBuffer_as_sequence = {
-    .sq_length = (lenfunc)BamBlockBuffer__len__
-};
 
 PyDoc_STRVAR(BamBlockBuffer_write_doc,
 "Write a BamRecord object into the BamBlockBuffer.\n"
@@ -537,7 +529,6 @@ static PyTypeObject BamBlockBuffer_type = {
     .tp_dealloc = BamBlockBuffer_dealloc,
     .tp_init = BamBlockBuffer__init__, 
     .tp_new = PyType_GenericNew,
-    .tp_as_sequence = &BamBlockBuffer_as_sequence,
     .tp_doc = BamBlockBuffer__doc__,
     .tp_members = BamBlockBuffer_members,
     .tp_methods = BamBlockBuffer_methods,