Faster writing than pysam now! Yeah! :tada:

### Checklist
- [ ] Pull request details were added to CHANGELOG.rst
- [ ] Documentation was updated (if needed)