@rhpvorderman 2022-02-24
Before:
```
Benchmark #1: python benchmarks/bam_read_and_write_uncompressed.py ~/test/test.uncompressed.bam /dev/null
  Time (mean ± σ):      2.949 s ±  0.042 s    [User: 2.796 s, System: 0.153 s]
  Range (min … max):    2.903 s …  3.037 s    10 runs
```

After:
```
Benchmark #1: python benchmarks/bam_read_and_write_uncompressed.py ~/test/test.uncompressed.bam /dev/null
  Time (mean ± σ):      2.084 s ±  0.038 s    [User: 1.939 s, System: 0.145 s]
  Range (min … max):    2.037 s …  2.161 s    10 runs
```

Read speed is at 950 ms for uncompressed bam. That means going from 1.9 to 1.15 seconds for writing. A 40% reduction in time. 
Read speed here:
```
Benchmark #1: python benchmarks/bam_read.py ~/test/test.uncompressed.bam
  Time (mean ± σ):     946.5 ms ±  20.0 ms    [User: 807.1 ms, System: 139.3 ms]
  Range (min … max):   918.7 ms … 985.3 ms    10 runs
 ```

Pysam (conda version 0.18.0) speed here:
```
Benchmark #1: python benchmarks/pysam_read_and_write_uncompressed.py ~/test/test.uncompressed.bam /dev/null
  Time (mean ± σ):      2.829 s ±  0.025 s    [User: 2.690 s, System: 0.139 s]
  Range (min … max):    2.796 s …  2.876 s    10 runs
```
Pysam read speed:
```
Benchmark #1: python benchmarks/pysam_read.py ~/test/test.uncompressed.bam /dev/null
  Time (mean ± σ):      1.429 s ±  0.030 s    [User: 1.285 s, System: 0.143 s]
  Range (min … max):    1.397 s …  1.505 s    10 runs
 ```